refactor: DetectEmotion services 디렉토리로 이동 playlist api 임의 값 수정
